move SynchronizedTypeLoader to a non-inner
