+++ src/main/java/cuchaz/enigma/Deobfuscator.java
@@ -15,7 +15,6 @@ import com.google.common.base.Stopwatch;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
-import com.strobel.assembler.metadata.Buffer;
 import com.strobel.assembler.metadata.MetadataSystem;
 import com.strobel.assembler.metadata.TypeDefinition;
 import com.strobel.assembler.metadata.TypeReference;
@@ -667,31 +666,4 @@ public class Deobfuscator {
 		String transform(ClassNode node, ClassWriter writer);
 	}
 
-	private class SynchronizedTypeLoader implements ITranslatingTypeLoader {
-		private final TranslatingTypeLoader delegate;
-
-		private SynchronizedTypeLoader(TranslatingTypeLoader delegate) {
-			this.delegate = delegate;
-		}
-
-		@Override
-		public List<String> getClassNamesToTry(String className) {
-			return delegate.getClassNamesToTry(className);
-		}
-
-		@Override
-		public List<String> getClassNamesToTry(ClassEntry obfClassEntry) {
-			return delegate.getClassNamesToTry(obfClassEntry);
-		}
-
-		@Override
-		public String transformInto(ClassNode node, ClassWriter writer) {
-			return delegate.transformInto(node, writer);
-		}
-
-		@Override
-		public synchronized boolean tryLoadType(String internalName, Buffer buffer) {
-			return delegate.tryLoadType(internalName, buffer);
-		}
-	}
 }
